% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemming.R
\name{add_stem}
\alias{add_stem}
\title{Add a single token-stem mapping}
\usage{
add_stem(token, stem)
}
\arguments{
\item{token}{A new token}

\item{stem}{Its stem}
}
\value{
None
}
\description{
This overwrites or augments the package's existing token to stem mapping.
}
\examples{
add_stem("zürriyetindensin", "zürriyet")

}
