% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemming.R
\name{add_stems}
\alias{add_stems}
\title{Add a file of new token-stems pairs}
\usage{
add_stems(fname)
}
\arguments{
\item{fname}{Name of the file}
}
\value{
Nothing
}
\description{
The file should have lines of the form <token>[tab]<stem>. Empty lines and
lines starting with # will be ignored.  An example of the correct format
is \code{system.file("extdata", "manual.dict", package = "Resha")}
}
\details{
Note that changes to the token-stem mapping are not maintained
between package loadings.
}
\examples{
newfile <- system.file("extdata", "manual.dict", package = "Resha")
add_stems(newfile)

}

