% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resha.R
\docType{package}
\name{Resha}
\alias{Resha}
\alias{Resha-package}
\title{Resha: A package for stemming Turkish less vigorously than Snowball}
\description{
This package is an R version of Harun Reşit Zafer's resha stemmer for
Turkish, from which it takes its language resources.
}
\details{
The main function is \link{\code{stem}} which take a vector of tokens
and returns for each, either its stem from a look-up table, or if none
can be found the original token.  Tokens are truncated up to the first
apostrophe, if any, before stemming.

You can add your own token-stem mappings either individually using
\link{\code{add_stem}} or as together using an input file using
\link{\code{add_stems}}.
}

