% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemming.R
\name{wordStem}
\alias{wordStem}
\title{Stem Turkish tokens}
\usage{
wordStem(x, ...)
}
\arguments{
\item{x}{A token or a vector of tokens}

\item{...}{Extra arguments, currently ignored}
}
\value{
A stemmed token or vector of stemmed tokens, or the originals if no stems could be found
}
\description{
This function attempts to stem Turkish tokens using a look-up table (derived
from Nuve) as a fast substitute for more complex but more accurate
morphological analysis.  If tokens contain an apostrophe, only characters
before are stemmed and the remainder discarded.
}
\details{
This code should work the same way as the original Java implementation.
The interface on the other hand is designed to work feel like
the \code{SnowballC} package.

Note: This function assumes that all proper nouns are capitalized and
other words are not.  (The stemmer is built around a look up table) so
you may wish to check that non-proper nouns starting sentences are
lowercased appropriately in the input.
}
\examples{
  toks <- c("kitapçığında", "kitapçıdaki", "İstanbul'da")
  wordStem(toks)

}
\references{
\itemize{
  \item{Resha}{\url{https://github.com/hrzafer/resha-turkish-stemmer}},
  \item{Nuve}{\url{https://github.com/hrzafer/nuve}}
}
}
